fix: 修复保持最后输入模式未按预期切换的问题
