+++ Fire/FireInputController.swift
@@ -17,12 +17,12 @@ typealias NotificationObserver = (name: Notification.Name, callback: (_ notifica
 class FireInputController: IMKInputController {
     private var _candidates: [Candidate] = []
     private var _hasNext: Bool = false
-    private var inputMode: InputMode {
+    internal var inputMode: InputMode {
         get { Fire.shared.inputMode }
         set(value) { Fire.shared.inputMode = value }
     }
 
-    private var temp: (
+    internal var temp: (
         observerList: [NSObjectProtocol],
         monitorList: [Any?]
     ) = (
@@ -361,77 +361,5 @@ class FireInputController: IMKInputController {
                 }
             })
         ]
-    }
-
-    /**
-     * 根据当前输入的应用改变输入模式
-     */
-    private func activeClientInputMode() {
-        guard let identifier = client()?.bundleIdentifier() else { return }
-        if let appSetting = Defaults[.appSettings][identifier],
-           let mode = InputMode(rawValue: appSetting.inputModeSetting.rawValue) {
-            print("[FireInputController] activeClientInputMode from setting : \(identifier), \(mode)")
-            Fire.shared.toggleInputMode(mode)
-            return
-        }
-        if !Defaults[.keepAppInputMode] { return }
-        // 启用APP缓存设置
-        if let appSetting = Fire.shared.appSettingCache.get(bundleIdentifier: identifier),
-           let mode = InputMode(rawValue: appSetting.inputModeSetting.rawValue) {
-            print("[FireInputController] activeClientInputMode from cache: \(identifier), \(mode)")
-            Fire.shared.toggleInputMode(mode)
-        }
-    }
-
-    /**
-     * 由于使用recognizedEvents在一些场景下不能监听到flagChanged事件，比如保存文件场景
-     * 所以这里需要使用NSEvent.addGlobalMonitorForEvents监听shift键被按下
-     */
-    override func activateServer(_ sender: Any!) {
-        NSLog("[FireInputController] activate server: \(client()?.bundleIdentifier() ?? sender.debugDescription)")
-
-        // 监听candidateView点击，翻页事件
-        notificationList().forEach { (observer) in temp.observerList.append(NotificationCenter.default.addObserver(
-            forName: observer.name, object: nil, queue: nil, using: observer.callback
-        ))}
-        if Defaults[.disableEnMode] {
-            Fire.shared.toggleInputMode(.zhhans)
-            return
-        }
-
-        activeClientInputMode()
-        temp.monitorList.append(NSEvent.addGlobalMonitorForEvents(matching: .flagsChanged) { (event) in
-            if !InputSource.shared.isSelected() {
-                return self.clearEventListener()
-            }
-            _ = self.handle(event, client: self.client())
-        })
-    }
-    func clearEventListener() {
-        temp.monitorList.forEach { (monitor) in
-            if let m = monitor {
-                NSEvent.removeMonitor(m)
-            }
-        }
-        temp.observerList.forEach { (observer) in
-            NotificationCenter.default.removeObserver(observer)
-        }
-        temp.monitorList = []
-        temp.observerList = []
-    }
-    override func deactivateServer(_ sender: Any!) {
-        NSLog("[FireInputController] deactivate server: \(client()?.bundleIdentifier() ?? "no client deactivate")")
-        if let identifier = client()?.bundleIdentifier(), !Defaults[.disableEnMode] {
-        // 缓存当前输入模式
-            Fire.shared.appSettingCache.add(
-                bundleIdentifier: identifier,
-                setting: ApplicationSettingItem(
-                    bundleId: identifier,
-                    inputMs: InputModeSetting(rawValue: inputMode.rawValue)!
-                )
-            )
-        }
-        clean()
-        clearEventListener()
     }
 }
